Back to uint64_t

+++ test/t8_cmesh/t8_gtest_compute_first_element.cxx
@@ -135,8 +135,8 @@ TEST_P (t8_gtest_rank_times_global_num_elems_over_size, large_numbers)
    * We use integer division, therefore we store the remainder of each update to
    * prevent rounding errors.
   */
-  uint32_t size = 1;
-  // uint64_t size = 1;
+  // uint32_t size = 1;
+  uint64_t size = 1;
   for (uint32_t isize = 1; isize < size_iter; ++isize) {
     /* The very first result is 1 * 1 / size */
     uint64_t check_result_elem = 1 / size;
@@ -177,7 +177,7 @@ TEST_P (t8_gtest_rank_times_global_num_elems_over_size, large_numbers)
         ASSERT_EQ (computed_result, check_result)
           << "rank: " << rank << " num_elems: " << num_elems << " size: " << size;
 
-        // /* Update the result with respect to the updated rank */
+        /* Update the result with respect to the updated rank */
         check_result *= rank_growth;
         check_result += rank_growth * rank_remainder / size;
         rank_remainder = (rank_growth * rank_remainder) % size;
